03-Orchestration/requirements.txt
apache-airflow==3.0.0
apache-airflow-providers-postgres   

# Core Data Science / ML dependencies
pandas
scikit-learn
xgboost
mlflow
pyarrow   
